requirements.txt
torch
numpy
Pillow
torchvision
opencv-python